some channel makers -> producer makers
