+++ Sources/Channel_makers.swift
@@ -160,49 +160,6 @@ public func channel<C: Collection, Success>(
 }
 
 // MARK: -
-/// Convenience constructor of Channel
-/// Gives an access to an underlying Producer to a provided block
-public func channel<Update, Success>(
-  executor: Executor = .immediate,
-  cancellationToken: CancellationToken? = nil,
-  bufferSize: Int = AsyncNinjaConstants.defaultChannelBufferSize,
-  block: @escaping (_ producer: Producer<Update, Success>) throws -> Void
-  ) -> Channel<Update, Success> {
-  // TEST: ChannelMakersTests.testMakeChannelWithProducerProvidingBlock
-
-  let producer = Producer<Update, Success>(bufferSize: bufferSize)
-  cancellationToken?.add(cancellable: producer)
-  executor.execute(from: nil) { (originalExecutor) in
-    do {
-      try block(producer)
-    } catch {
-      producer.fail(error, from: originalExecutor)
-    }
-  }
-  return producer
-}
-
-/// Convenience constructor of Channel
-/// Gives an access to an underlying Producer to a provided block
-public func channel<C: ExecutionContext, Update, Success>(
-  context: C,
-  executor: Executor? = nil,
-  cancellationToken: CancellationToken? = nil,
-  bufferSize: Int = AsyncNinjaConstants.defaultChannelBufferSize,
-  block: @escaping (_ context: C, _ strongProducer: Producer<Update, Success>) throws -> Void
-  ) -> Channel<Update, Success> {
-  // TEST: ChannelMakersTests.testMakeChannelWithProducerProvidingBlockWithDeadContext
-
-  return channel(executor: executor ?? context.executor,
-                 cancellationToken: cancellationToken,
-                 bufferSize: bufferSize
-  ) { (producer: Producer<Update, Success>) in
-    context.addDependent(completable: producer)
-    try block(context, producer)
-  }
-}
-
-// MARK: -
 
 /// Convenience shortcuts for making completed channel
 public extension Channel {
